[bugfix] make unbound_version `--check` safe, update and fix README
the previoius implementation does not work with `ansible-play --check`.
it worked only after unbound is installed, i.e. it did not when the play
with `--check` is run for the first time because the package has not yet
installed.

introduce `unbound_package` variable.

also, remove chroot support, which was removed in past release,
documentation from README.